Basic printf() implementation as xprintf() macro.
The arguments to the macro are the same as the arguments to the
function. A macro has to be used in order to call TWIG_MKSTR although I
probably ought to investigate if there's a way to make a static inline
version.

Slightly rough around the edges: the interface to printf has a
global private variable which is used to hold the parsing state and the
packer context, so that those do not have to be passed to the xprintf()
macro - thus making it more printf like. Ideally, an application could
"opt out" of that functionality, allowing there to be no hidden state
(and much easier thread safety)