Commonize printf formatter w/o conditional compile
Trying to completely commonize the printf formatter turned out to be a
fruitless exercise. In order to put the majority of the source code
under common/, several large chunks of the code would need to be
ifdef'd in a way that makes the code complicated to read, and
ultimately, not actually common code. The testability benefit in
shared code would be dashed by the maintainability headache of
conditional compilation.

The resulting architecture contains actually fully common code in
common/, and decoder / encoder specific code which mirror each other
except for a slight change in arguments in their respective directories.
The common code is built separately for encoder and decoder, by only
collecting files by their function in common/ and not actually building
them.