Decoder and Encoder both compile & run
Small victories... everything is *compiling* and running without
immediate crash, despite the somewhat messy build & code layout.

The decoder does very little but a proof-of-concept: it prints out the
string table & offsets read from the various binary files, which have to
be generated first by posix-extractor.sh. It then parses the index (but
not arguments) of a message in a file stream.

Enough is complete at this point to say the architecture of twig is
solid & can work in the general case.