choosing method 1 for unpacker
now I need to deal with the type-handling mess, because we need to know
not only size, but sign for integers.

partially dealt w/ the buffering issue, dynabuf is added but not handled
yet. really only for strings which could be > 256 bytes (size of the
temp buf read in by unpackarg). For ints and floats we can be sure that
they will decode okay even if they are split across two read() calls
because they will fit in the buffer.

notes in the source code declare what is yet to be done.